Moved signing ethereum command to a dedicated file
We want to enhance the code for ethereum signing so it may be a good
idea to extract these to a separate file.
